requirements.txt
appnope==0.1.3
asttokens==2.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
fonttools==4.38.0
idna==3.4
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.1
Js2Py==0.74
jupyter_client==7.4.6
jupyter_core==5.0.0
kiwisolver==1.4.4
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.4
packaging==21.3
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pipwin==0.5.2
platformdirs==2.5.4
prompt-toolkit==3.0.32
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyjsparser==2.7.1
pyparsing==3.0.9
PyPrind==2.11.3
pyproj==3.4.0
pyshp==2.3.1
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
Shapely==1.8.5.post1
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.6.1
tornado==6.2
traitlets==5.5.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
wcwidth==0.2.5
